moving class definitions to source files
