+++ vowpalwabbit/cb_explore_adf_bag.cc
@@ -19,6 +19,31 @@ namespace cb_explore_adf
 {
 namespace bag
 {
+struct cb_explore_adf_bag : public cb_explore_adf_base
+{
+ private:
+  float _epsilon;
+  size_t _bag_size;
+  bool _greedify;
+  bool _first_only;
+  std::shared_ptr<rand_state> _random_state;
+
+  v_array<ACTION_SCORE::action_score> _action_probs;
+  std::vector<float> _scores;
+  std::vector<float> _top_actions;
+
+ public:
+  cb_explore_adf_bag(
+      float epsilon, size_t bag_size, bool greedify, bool first_only, std::shared_ptr<rand_state> random_state);
+  template <bool is_learn>
+  static void predict_or_learn(cb_explore_adf_bag& data, LEARNER::multi_learner& base, multi_ex& examples);
+  ~cb_explore_adf_bag();
+
+ private:
+  template <bool is_learn>
+  void predict_or_learn_impl(LEARNER::multi_learner& base, multi_ex& examples);
+};
+
 cb_explore_adf_bag::cb_explore_adf_bag(
     float epsilon, size_t bag_size, bool greedify, bool first_only, std::shared_ptr<rand_state> random_state)
     : _epsilon(epsilon), _bag_size(bag_size), _greedify(greedify), _random_state(random_state)
