+++ vowpalwabbit/cb_explore_adf_cover.cc
@@ -19,6 +19,36 @@ namespace cb_explore_adf
 {
 namespace cover
 {
+struct cb_explore_adf_cover : public cb_explore_adf_base
+{
+ private:
+  size_t _cover_size;
+  float _psi;
+  bool _nounif;
+  bool _first_only;
+  size_t _counter;
+  LEARNER::multi_learner* _cs_ldf_learner;
+  GEN_CS::cb_to_cs_adf _gen_cs;
+
+  v_array<ACTION_SCORE::action_score> _action_probs;
+  std::vector<float> _scores;
+  COST_SENSITIVE::label _cs_labels;
+  COST_SENSITIVE::label _cs_labels_2;
+  v_array<COST_SENSITIVE::label> _prepped_cs_labels;
+  v_array<CB::label> _cb_labels;
+
+ public:
+  cb_explore_adf_cover(size_t cover_size, float psi, bool nounif, bool first_only,
+      LEARNER::multi_learner* cs_ldf_learner, LEARNER::single_learner* scorer, size_t cb_type);
+  template <bool is_learn>
+  static void predict_or_learn(cb_explore_adf_cover& data, LEARNER::multi_learner& base, multi_ex& examples);
+  ~cb_explore_adf_cover();
+
+ private:
+  template <bool is_learn>
+  void predict_or_learn_impl(LEARNER::multi_learner& base, multi_ex& examples);
+};
+
 cb_explore_adf_cover::cb_explore_adf_cover(size_t cover_size, float psi, bool nounif, bool first_only,
     LEARNER::multi_learner* cs_ldf_learner, LEARNER::single_learner* scorer, size_t cb_type)
     : _cover_size(cover_size), _psi(psi), _nounif(nounif), _first_only(first_only), _cs_ldf_learner(cs_ldf_learner)
