+++ vowpalwabbit/cb_explore_adf_first.cc
@@ -18,6 +18,23 @@ namespace cb_explore_adf
 {
 namespace first
 {
+struct cb_explore_adf_first : public cb_explore_adf_base
+{
+ private:
+  size_t _tau;
+  float _epsilon;
+
+ public:
+  cb_explore_adf_first(size_t tau, float epsilon);
+  template <bool is_learn>
+  static void predict_or_learn(cb_explore_adf_first& data, LEARNER::multi_learner& base, multi_ex& examples);
+  ~cb_explore_adf_first() = default;
+
+ private:
+  template <bool is_learn>
+  void predict_or_learn_impl(LEARNER::multi_learner& base, multi_ex& examples);
+};
+
 cb_explore_adf_first::cb_explore_adf_first(size_t tau, float epsilon)
   : _tau(tau), _epsilon(epsilon) {}
