+++ vowpalwabbit/cb_explore_adf_softmax.h
@@ -8,8 +8,6 @@ license as described in the file LICENSE.
 #include "cb_explore_adf_common.h"
 #include "reductions_fwd.h"
 
-#include <vector>
-
 namespace VW
 {
 namespace cb_explore_adf
@@ -17,23 +15,6 @@ namespace cb_explore_adf
 namespace softmax
 {
 LEARNER::base_learner* setup(VW::config::options_i& options, vw& all);
-
-struct cb_explore_adf_softmax : public cb_explore_adf_base
-{
- private:
-  float _epsilon;
-  float _lambda;
-
- public:
-  cb_explore_adf_softmax(float epsilon, float lambda);
-  template <bool is_learn>
-  static void predict_or_learn(cb_explore_adf_softmax& data, LEARNER::multi_learner& base, multi_ex& examples);
-  ~cb_explore_adf_softmax() = default;
-
- private:
-  template <bool is_learn>
-  void predict_or_learn_impl(LEARNER::multi_learner& base, multi_ex& examples);
-};
 }  // namespace softmax
 }  // namespace cb_explore_adf
 }  // namespace VW
